# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mafia World Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gangs:
    get:
      operationId: listGangs
      tags: [gangs]
      summary: List all gangs
      responses:
        "200":
          description: List of gangs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gang"
    post:
      operationId: createGang
      tags: [gangs]
      summary: Create a new gang
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GangInput"
      responses:
        "201":
          description: Created gang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gang"
  /gangs/{gangId}:
    get:
      operationId: getGang
      tags: [gangs]
      summary: Get gang details
      parameters:
        - in: path
          name: gangId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gang"
  /gangs/{gangId}/join:
    post:
      operationId: joinGang
      tags: [gangs]
      summary: Join a gang
      parameters:
        - in: path
          name: gangId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GangMember"
...
  /gangs/{gangId}/members:
    get:
      operationId: listGangMembers
      tags: [gangs]
      summary: List gang members
      parameters:
        - in: path
          name: gangId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GangMember"
  /gangs/{gangId}/members/{memberId}/promote:
    post:
      operationId: promoteMember
      tags: [gangs]
      summary: Promote a gang member
      parameters:
        - in: path
          name: gangId
          required: true
          schema:
            type: integer
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteMemberInput"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GangMember"
  /gangs/{gangId}/deposit:
    post:
      operationId: depositToGangTreasury
      tags: [gangs]
      summary: Deposit money to gang treasury
      parameters:
        - in: path
          name: gangId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreasuryDeposit"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- WEAPONS ---
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Gang:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        treasury:
          type: integer
        memberCount:
          type: integer
        bossName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - treasury
        - memberCount
        - bossName
        - createdAt
    GangInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
      required:
        - name
        - description
    GangMember:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        username:
          type: string
        level:
          type: integer
        rank:
          type: string
          enum: [Soldier, Capo, Underboss, Consigliere, Boss]
        joinedAt:
          type: string
      required:
        - id
        - playerId
        - username
        - level
        - rank
        - joinedAt
    PromoteMemberInput:
      type: object
      properties:
        rank:
          type: string
          enum: [Soldier, Capo, Underboss, Consigliere, Boss]
      required:
        - rank
    TreasuryDeposit:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
      required:
        - amount
...